pyproject.toml
[project]
name = "datasetpro"
version = "0.1.0"
description = "Pipeline de construcción de dataset para DeepShield"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.33.0",
    "albumentations>=1.4.0",
    "diffusers>=0.30.0",
    "h5py>=3.11.0",
    "insightface>=0.7.3",
    "numpy>=1.26.0",
    "onnxruntime-gpu>=1.19.0",
    "opencv-python-headless>=4.10.0.84",
    "pillow>=10.4.0",
    "scikit-learn>=1.5.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "tqdm>=4.66.0",
    "transformers>=4.44.0",
]